% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equipment.R
\name{Equipment}
\alias{Equipment}
\title{R6 Class for recipe equipment}
\description{
An \code{Equipment} object is an R6 object, and it can be created using
\code{Equipment$new()}. A \code{name} field for the equipment piece is required, and a
\code{quantity} field can be added if desired.
}
\section{Fields}{

\describe{
\item{\code{name}}{Scalar character vector; the name of the piece of equipment.}

\item{\code{quantity}}{Scalar numeric vector; the number of equipment units needed. Defaults to
1.}
}}

\seealso{
\link{Recipe}

Other recipe components: 
\code{\link{Image}},
\code{\link{Ingredient}},
\code{\link{Instruction}}
}
\concept{recipe components}
