% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Instruction.R
\name{Instruction}
\alias{Instruction}
\title{R6 Class for recipe instructions}
\description{
R6 Class for recipe instructions
}
\section{Creating the instruction object}{

An \code{Instruction} is an R6 object, and it can be created using
\code{Instruction$new()}. It only has one element, \code{text}, that contains the
instruction detail.

\describe{
\item{text}{
Scalar character vector; the wording of one step of a recipe.
}
}
}

