% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equipment.R
\name{Equipment}
\alias{Equipment}
\title{R6 Class for recipe equipment}
\description{
R6 Class for recipe equipment
}
\section{Creating the equipment object}{

An \code{Equipment} object is an R6 object, and it can be created using
\code{Equipment$new()}. The name of the equipment piece is required, and the
quantity can be omitted if desired.

\describe{
\item{name}{
Scalar character vector; the name of the piece of equipment.
}
\item{quantity}{
Scalar numeric vector; the number of equipment units needed. Defaults to
1.
}
}
}

