% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipe.R
\name{Recipe}
\alias{Recipe}
\title{R6 Class providing standard format for recipes}
\description{
This class will encapsulate various recipe elements into a standard,
consistent format. Once the object is built, it will be easier to work with
when incorporating in other projects, i.e., a website.
}
\section{Fields}{

\describe{
\item{\code{name}}{Scalar character vector; the name of the recipe.}

\item{\code{summary}}{Character vector; a brief description of what the recipe makes.}

\item{\code{servings}}{Numeric vector of length 1 or 2; if only one value provided, the number of
servings the recipe produces. If 2 values provided, the minimum and maximum
number of servings that the recipe produces.}

\item{\code{time}}{(Optionally named) list of \link[lubridate:duration]{lubridate::duration} objects (or objects that
can be converted to \code{lubridate::duration} objects, i.e., the total number
of seconds (since \code{lubridate::duration} objects are measured in seconds)
or string representations of durations, e.g., "14 hours 20 minutes"); you
can provide a list with only one element to represent the total recipe
time, or you can provide a named list with elements that represent
different time components, e.g., "active", "prep", etc.. Providing an
unnamed lists with multiple elements will return an error because it is
unclear which element corresponds to which time component.}

\item{\code{images}}{A list of \link{Image} objects associated with this \code{Recipe}.}

\item{\code{ingredients}}{A list of \link{Ingredient} objects associated with this \code{Recipe}.}

\item{\code{equipment}}{A list of \link{Equipment} objects associated with this \code{Recipe}.}

\item{\code{instructions}}{A list of \link{Instruction} objects associated with this \code{Recipe}.}
}}

\section{Creating the recipe object}{

A \code{Recipe} is an R6 object, and it can be created using \code{Recipe$new()}. A
\code{Recipe} has several different fields that you could pass as arguments
when initializing, however, you only need to provide the \code{name} when first
initializing; other fields can be added at a later time.
}

\section{Uses for the recipe object}{

This class was created mainly to store recipes in a consistent format.
\code{Recipe} fields can be manually provided, but I've been webscraping and
adding fields that way. Once fully built, different recipe fields can be
accessed using the \code{$} operator.
}

\examples{
# An example recipe for spaghetti sauce:
r <- Recipe$new('Spaghetti Sauce')
print(r)

# Add a summary:
r$summary <- paste(
  'This is my grandmother\'s secret recipe for her award-winning spaghetti',
  'sauce. It has been passed down for generations and dates back to the',
  '1700\'s!'
)
print(r)

# Other elements are additional R6 classes. See the help page of each for
# additional information.

}
\seealso{
\link{Image} \link{Ingredient} \link{Equipment} \link{Instruction} \link[lubridate:duration]{lubridate::duration}
}
