% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recipe.R
\name{Recipe}
\alias{Recipe}
\title{R6 Class providing standard format for recipes}
\description{
This class will encapsulate various recipe elements into a standard,
consistent format. Once the object is built, it will be easier to work with
when incorporating in other projects, i.e., a website.
}
\section{Fields}{

\describe{
\item{\code{name}}{Scalar character vector; the name of the recipe.}

\item{\code{summary}}{Character vector; a brief description of what the recipe makes.}

\item{\code{images}}{A list of \link{Image} objects associated with this \code{Recipe}.}

\item{\code{ingredients}}{A list of \link{Ingredient} objects associated with this \code{Recipe}.}

\item{\code{equipment}}{A list of \link{Equipment} objects associated with this \code{Recipe}.}

\item{\code{instructions}}{A list of \link{Instruction} objects associated with this \code{Recipe}.}

\item{\code{servings}}{Scalar numeric vector; the number of serving the recipe will produce.}
}}

\section{Creating the recipe object}{

A \code{Recipe} is an R6 object, and it can be created using \code{Recipe$new()}. A
\code{Recipe} has several different fields that you could pass as arguments
when initializing, however, you only need to provide the \code{name} when first
initializing; other fields can be added at a later time.
}

\section{Uses for the recipe object}{

This class was created mainly to store recipes in a consistent format.
\code{Recipe} fields can be manually provided, but I've been webscraping and
adding fields that way. Once fully built, different recipe fields can be
accessed using the \code{$} operator.
}

\examples{
# An example recipe for spaghetti sauce:
r <- Recipe$new('Spaghetti Sauce')
print(r)

# Add a summary:
r$summary <- paste(
  'This is my grandmother\'s secret recipe for her award-winning spaghetti',
  'sauce. It has been passed down for generations and dates back to the',
  '1700\'s!'
)
print(r)

# Other elements are additional R6 classes. See the help page of each for
# additional information.

}
\seealso{
\link{Equipment} \link{Instruction} \link{Image} \link{Ingredient}
}
